refactor TeamScoreCumulative to its own file
also renames MatchCumulativeScoreCounter to TeamScoreCumulative
