+++ osu.Game.Tournament/Screens/Gameplay/Components/TeamDisplay.cs
@@ -1,13 +1,9 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Game.Graphics;
-using osu.Game.Graphics.Sprites;
-using osu.Game.Graphics.UserInterface;
 using osu.Game.Tournament.Components;
 using osu.Game.Tournament.Models;
 using osuTK;
@@ -16,39 +12,9 @@ namespace osu.Game.Tournament.Screens.Gameplay.Components
 {
     public partial class TeamDisplay : DrawableTournamentTeam
     {
-        protected partial class MatchCumulativeScoreCounter : CommaSeparatedScoreCounter
-        {
-            private OsuSpriteText displayedSpriteText = null!;
-            private const int font_size = 50;
-            private Bindable<bool> useCumulativeScore = null!;
-
-            [Resolved]
-            private LadderInfo ladder { get; set; } = null!;
-
-            public MatchCumulativeScoreCounter()
-            {
-                Margin = new MarginPadding(8);
-            }
-
-            protected override void LoadComplete()
-            {
-                base.LoadComplete();
-
-                useCumulativeScore = ladder.CumulativeScore.GetBoundCopy();
-                useCumulativeScore.BindValueChanged(v => displayedSpriteText.Alpha = v.NewValue ? 1 : 0, true);
-            }
-
-            protected override OsuSpriteText CreateSpriteText() => base.CreateSpriteText().With(s =>
-            {
-                displayedSpriteText = s;
-                displayedSpriteText.Spacing = new Vector2(-6);
-                displayedSpriteText.Font = OsuFont.Torus.With(weight: FontWeight.SemiBold, size: font_size, fixedWidth: true);
-            });
-        }
-
         private readonly TeamScore score;
 
-        private readonly MatchCumulativeScoreCounter cumulativeScoreCounter;
+        private readonly TeamScoreCumulative teamScoreCumulative;
 
         private readonly TournamentSpriteTextWithBackground teamNameText;
 
@@ -145,7 +111,7 @@ namespace osu.Game.Tournament.Screens.Gameplay.Components
                                     },
                                 }
                             },
-                            cumulativeScoreCounter = new MatchCumulativeScoreCounter
+                            teamScoreCumulative = new TeamScoreCumulative
                             {
                                 Origin = anchor,
                                 Anchor = anchor,
@@ -156,7 +122,7 @@ namespace osu.Game.Tournament.Screens.Gameplay.Components
             };
             currentTeamScore.BindValueChanged(_ =>
             {
-                cumulativeScoreCounter.Current.Value = currentTeamScore.Value ?? 0;
+                teamScoreCumulative.Current.Value = currentTeamScore.Value ?? 0;
             }, true);
         }
 
@@ -176,7 +142,7 @@ namespace osu.Game.Tournament.Screens.Gameplay.Components
         private void updateDisplay()
         {
             score.FadeTo(ShowScore ? 1 : 0, 200);
-            cumulativeScoreCounter.FadeTo(ShowScore ? 1 : 0, 200);
+            teamScoreCumulative.FadeTo(ShowScore ? 1 : 0, 200);
         }
     }
 }
